pyproject.toml
[tool.poetry]
name = "speedstats"
version = "0.1.0"
description = "Internet Speed Statistic"
authors = ["Searge"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
speedtest-cli = "^2.1.3"
rich = "^10.1"
urllib3 = "^1.26"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"

autopep8 = "^1.5.6"
mypy = "^0.812"
wemake-python-styleguide = "^0.15.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.nitpick]
style = [
  "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/flake8.toml",
  "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/isort.toml"
]